+++ wallarm/provider/resource_rule_bola_counter.go
@@ -26,7 +26,7 @@ func resourceWallarmBolaCounter() *schema.Resource {
 	return &schema.Resource{
 		CreateContext: resourceWallarmBolaCounterCreate,
 		ReadContext:   resourceWallarmBolaCounterRead,
-		DeleteContext: resourceWallarmBolaCounterDelete,
+		DeleteContext: resourcerule.CounterDelete("wallarm_rule_bola_counter"),
 		Importer: &schema.ResourceImporter{
 			StateContext: resourcerule.Import("bola_counter"),
 		},
@@ -82,24 +82,4 @@ func resourceWallarmBolaCounterRead(_ context.Context, d *schema.ResourceData, m
 		return diag.FromErr(err)
 	}
 	return diag.FromErr(resourcerule.Read(d, clientID, apiClient(m), resourcerule.ReadOptionWithAction))
-}
-
-func resourceWallarmBolaCounterDelete(_ context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
-	client := apiClient(m)
-	clientID, err := retrieveClientID(d, m)
-	if err != nil {
-		return diag.FromErr(err)
-	}
-	ruleID := d.Get("rule_id").(int)
-	h := &wallarm.HintDelete{
-		Filter: &wallarm.HintDeleteFilter{
-			Clientid: []int{clientID},
-			ID:       []int{ruleID},
-		},
-	}
-	if err := client.HintDelete(h); err != nil {
-		return diag.FromErr(err)
-	}
-
-	return nil
 }